requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
SQLAlchemy
flask-restful==0.3.9
flask-security-too
email_validator==1.3.0
bcrypt==4.0.1
pytest==7.2.0
coverage==7.0.1
beautifulsoup4==4.11.1
Flask-Migrate==4.0.0
redis
celery[redis]==5.2.3
flask-sse 
gunicorn
gevent
Flask-Caching